+++ src/test/java/com/jnape/palatable/lambda/functions/Fn3Test.java
@@ -3,6 +3,7 @@ package com.jnape.palatable.lambda.functions;
 import org.junit.Test;
 
 import static com.jnape.palatable.lambda.adt.hlist.HList.tuple;
+import static com.jnape.palatable.lambda.functions.Fn3.fn3;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
@@ -31,11 +32,13 @@ public class Fn3Test {
     }
 
     @Test
-    public void fn3() {
+    public void staticFactoryMethods() {
         Fn1<String, Fn2<String, String, String>> fn1 = a -> (b, c) -> a + b + c;
-        assertEquals("abc", Fn3.fn3(fn1).apply("a", "b", "c"));
+        assertEquals("abc", fn3(fn1).apply("a", "b", "c"));
 
         Fn2<String, String, Fn1<String, String>> fn2 = (a, b) -> c -> a + b + c;
-        assertEquals("abc", Fn3.fn3(fn2).apply("a", "b", "c"));
+        assertEquals("abc", fn3(fn2).apply("a", "b", "c"));
+
+        assertEquals("abc", Fn3.<String, String, String, String>fn3((a, b, c) -> a + b + c).apply("a", "b", "c"));
     }
 }