+++ src/test/java/com/jnape/palatable/lambda/functions/Fn4Test.java
@@ -3,6 +3,7 @@ package com.jnape.palatable.lambda.functions;
 import org.junit.Test;
 
 import static com.jnape.palatable.lambda.adt.hlist.HList.tuple;
+import static com.jnape.palatable.lambda.functions.Fn4.fn4;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
@@ -31,14 +32,16 @@ public class Fn4Test {
     }
 
     @Test
-    public void fn4() {
+    public void staticFactoryMethods() {
         Fn1<String, Fn3<String, String, String, String>> fn1 = a -> (b, c, d) -> a + b + c + d;
-        assertEquals("abcd", Fn4.fn4(fn1).apply("a", "b", "c", "d"));
+        assertEquals("abcd", fn4(fn1).apply("a", "b", "c", "d"));
 
         Fn2<String, String, Fn2<String, String, String>> fn2 = (a, b) -> (c, d) -> a + b + c + d;
-        assertEquals("abcd", Fn4.fn4(fn2).apply("a", "b", "c", "d"));
+        assertEquals("abcd", fn4(fn2).apply("a", "b", "c", "d"));
 
         Fn3<String, String, String, Fn1<String, String>> fn3 = (a, b, c) -> (d) -> a + b + c + d;
-        assertEquals("abcd", Fn4.fn4(fn3).apply("a", "b", "c", "d"));
+        assertEquals("abcd", fn4(fn3).apply("a", "b", "c", "d"));
+
+        assertEquals("abcd", Fn4.<String, String, String, String, String>fn4((a, b, c, d) -> a + b + c + d).apply("a", "b", "c", "d"));
     }
 }