+++ src/test/java/com/jnape/palatable/lambda/functions/Fn5Test.java
@@ -3,6 +3,7 @@ package com.jnape.palatable.lambda.functions;
 import org.junit.Test;
 
 import static com.jnape.palatable.lambda.adt.hlist.HList.tuple;
+import static com.jnape.palatable.lambda.functions.Fn5.fn5;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
@@ -32,17 +33,19 @@ public class Fn5Test {
     }
 
     @Test
-    public void fn5() {
+    public void staticFactoryMethods() {
         Fn1<String, Fn4<String, String, String, String, String>> fn1 = a -> (b, c, d, e) -> a + b + c + d + e;
-        assertEquals("abcde", Fn5.fn5(fn1).apply("a", "b", "c", "d", "e"));
+        assertEquals("abcde", fn5(fn1).apply("a", "b", "c", "d", "e"));
 
         Fn2<String, String, Fn3<String, String, String, String>> fn2 = (a, b) -> (c, d, e) -> a + b + c + d + e;
-        assertEquals("abcde", Fn5.fn5(fn2).apply("a", "b", "c", "d", "e"));
+        assertEquals("abcde", fn5(fn2).apply("a", "b", "c", "d", "e"));
 
         Fn3<String, String, String, Fn2<String, String, String>> fn3 = (a, b, c) -> (d, e) -> a + b + c + d + e;
-        assertEquals("abcde", Fn5.fn5(fn3).apply("a", "b", "c", "d", "e"));
+        assertEquals("abcde", fn5(fn3).apply("a", "b", "c", "d", "e"));
 
         Fn4<String, String, String, String, Fn1<String, String>> fn4 = (a, b, c, d) -> (e) -> a + b + c + d + e;
-        assertEquals("abcde", Fn5.fn5(fn4).apply("a", "b", "c", "d", "e"));
+        assertEquals("abcde", fn5(fn4).apply("a", "b", "c", "d", "e"));
+
+        assertEquals("abcde", Fn5.<String, String, String, String, String, String>fn5((a, b, c, d, e) -> a + b + c + d + e).apply("a", "b", "c", "d", "e"));
     }
 }