+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- `Fn3-8` static factory overloads to aid in coercing lambdas
 
 ## [3.0.3] - 2018-05-27
 ### Added